Extract tiles.rs too
